Mark grounded cells behind inactive floating cells as marine margin
Add logic to mark grounded cells in groundedMarineMarginMask if they
are behind inactive floating cells. Then the melt rate calculated from
those grounded cells is first applied to the thin inactive ice in front
of them. There is still a problem with some cells on the margins of
Petermann being marked as groundedMarineMarginMask when they should not
be, but the melt rate being applied is zero.