Enable COSP for RRTMGP
Enable COSP for RRTMGP. This requires modifying the get_cloud_optics
routines in cam_optics to pass snow optical properties so that these can
then be passed to the COSP driver. In doing so, we also pass liquid and
ice optical properties out of these routines so that we can output these
for diagnostic purposes.