Fix issue with unallocated elem structure in dynamics initialization
This commit addresses an issue with passing a null pointer as input to
a subroutine when the number of MPI ranks is greater than the number
of dynamics elements.

The element structure "elem" is unitialized on all MPI ranks that are
not tagged to solve dynamics.  Thus keeping the original declaration as
a null pointer.

In some cases with certain compiler flags active this can cause the
simulation to halt with an error related to passing the null pointer
"elem" as an input.

This fix allocates the elem pointer to be empty.

[BFB]