Update modules on Cori (cori-knl and cori-haswell) to reflect the current defaults. Except for the Intel compiler -- leaving this as-is at v18. We will update the compiler version in a different PR.
This PR updates
 cray-mpich/7.7.6 -> cray-mpich/7.7.10
 craype/2.5.18 -> craype/2.6.2
 cray-netcdf-hdf5parallel/4.6.1.3 -> cray-netcdf-hdf5parallel/4.6.3.2
 cray-hdf5-parallel/1.10.2.0 -> cray-hdf5-parallel/1.10.5.2
 cray-parallel-netcdf/1.8.1.4 -> cray-parallel-netcdf/1.11.1.1

These changes should be BFB and have been with testing so far (e3sm_developer plus maint-1.0 tests)
However, the IO-related module updates cause the memory use to increase (perhaps increasing buffers),
and this causes the comparing with baselines to fail.
Also, some of the tests are now faster and this causes compare tests to fail as the compute time was not expected.

[bfb]