Fix for LW cloud scattering and implementation of multiple iceflag options
A fix for the calculation of effective diameter (dei) to address
low-biased global LWCF to account for the difference between density of
ice crystals and density of solid ice. This is done by a multiplier of
rhoi/rhows in dei_grid = rei_grid * rhoi/rhows * 2._r8.

Three options for iceflag are now implemented. iceflag=3 is recommended
(default) to use MC6 ice cloud optics along with observed particle size
distribution. iceflag=1 is for MC6 ice cloud optics with 0.1 variance of
Gamma particle size distribbution.

[BFB] for default configuration, i.e., all umich flags are off.