Change MMF ERP test to ERS test
Change the MMF test in the integration suite from an ERP test to an ERS
test. ERP tests are currently failing the base restart comparison on
most machines, so in order to have a test in the integration suite that
will pass so that we can regression test we are temporarily changing to
ERS while we further investigate the ERP fail. Note that the ERP
actually passes on cori-knl, which is why we did not catch this sooner
as we had been doing all of our MMF testing there.