Enable building sam++ instead of sam for MMF
Enable building the sam++ code instead of sam for MMF. This adds two new
compsets, F-MMFXX and F-MMFXX-TEST that use sam++ as the source code for
the CRM. This is accomplished by adding a `-crm` option to the
atmosphere model configure script, whose argument should be the name of
the CRM code to build (currently either sam or samxx). Also some
cleaning up of config_component.xml to allow this to happen. This
*should* still allow building with sam for F-MMF[1,2] configurations,
but this needs to be tested. I believe the changes in `crm_physics.F90`
to accommodate sam++ might break the old sam build (arguments to `crm()`
are different).