SCREAM: avoid using GIT_BRANCH env var in testing util
If we are in a detached HEAD state, we can handle it.
We can merge origin/master into a detached head,
and we can check it out too. But if we use GIT_BRANCH,
then we make our utilities think that such branch locally
exists, which is not true. Some old versions of git cannot
check out , if it is the name of a remote branch that
is not fully qualified. E.g.

  git checkout my_branch

doesn't always work to check out origin/my_branch. In particular,
some (old) git versions would complain that my_branch is not
a valid ref.

Notice, however, that the fully qualified name always work.
That is,

  git checkout origin/my_branch

works also for older git versions (and will create a detached
head state).