Eliminate compile-time dependence on pcols in chemistry
The data type wetdep_inputs_t in chemistry/aerosol/wetdep.F90
refers to pcols in dimensioning some of the fields in the type.
These fields are changed to allocatable, and these fields
for variables of type wetdep_inputs_t are then allocated
in wetdep_inputs_set, and deallocated in a new subroutine
wetdep_inputs_unset. A call to wetdep_inputs_unset is added
to aero_model_wetdep in chemistry/bulk_aero/aero_model.F90
and to aero_model_wetdep in chemistry/modal_aero/aero_model.F90.

This removes one dependency on the compile-time specification
of pcols.

[BFB]