Implement pcols as a runtime option in EAM
In the earlier 2000's, empirical experiments indicated that setting
the first dimension of the chunk data structure at compile-time
allowed the compiler to generate more performance-efficient code, and
this motivated this design choice in CAM. However, CAM/EAM has evolved
significantly since then, and, for example, physics_types now uses
psetcols to define this first dimension, and the compiler may not
recognize that some fields have the first dimension set at
compile-time. Based on extensive experiments, compile-time setting
of PCOLS is no longer as big a performance advantage as previously
(though it is still somewhat faster). Being able to set pcols
at runtime increases ease of use, and increased flexibility
for future developments where pcols might need to be set after runtime
case specifications are read in.

Here pcols is implemented as a runtime option, currently specified
via the user_nl_cam namelist variable phys_chnk_fdim. phys_chnk_fdim
must be at least as large as 1, but there is no specified upper
bound. The current compile-time value of PCOLS is used to define the
default value for phys_chnk_fdim (and pcols). PCOLS is now
represented in EAM as the parameter ppcols, but is only referenced
in defining the default for phys_chnk_fdim.

The ability to set pcols at compile-time is retained as well.
This allows continued comparisons of performance between
compile-time and runtime pcols, and also to specify pcols at
compile-time for maximum performance for production jobs.

To enable compile-time setting of pcols, add '-DPPCOLS' to
CPPFLAGS in Macros.cmake for a case before building. With this
change, phys_chnk_fdim will be ignored (but a warning message
will be output if phys_chnk_fdim differs from the compile-time
setting of pcols.

[BFB]