Change resolution around Kamchatka Peninsula
Two different transition masks are defined, one for east of the
peninsula and one for west of it, with a transition right down
the center of the peninsula.  A key to a smooth transition is that
the shapes defining the "east" and "west" regions have the same
behavior south of the southern tip of the Peninsula.