Back out change in diag_surf
After re-examination with the total package of the changes,
the line

    call outfld('QFLX',     cam_in%cflx(1,1), pcols, lchnk)

in subroutine diag_surf in physics/cam/cam_diagnostics.F90
no longer fails to compile when pcols is not a parameter. As
clfx is a 2D array (pcols,pcnst), the original form of the
call to outfld is more self-documenting, and also matches the
style of other calls to outfld accessing slices of
multidimensional arrays. Here, the original code is restored.

[BFB]