SCREAM: changed name of github label that causes AT to return 'pass' immediately
The old label (AT: SKIP TESTING) might have been confused with what
the wip label (AT: WIP) means, namely to simply skip the PR.
Instead, this new label is meant to skip the tests, and immediately
report a 'pass'. The new label hopefully conveys the meanign better.

We also use a different prefix (CI) and not all uppercase after
the colon, so this helps differentiate between the AT built-in labels,
and the ones WE use to trigger special behavior in our testing.

The new label to make the AT skip tests is

CI: Integrate Without Testing