Reorder init. logic to allow pcols to be set in phys_grid_init
When set at runtime, pcols is set in phys_grid_setops, called
from within read_namelist, which is called from atm_init_mct.
In order to calculate a default runtime pcols that takes into
account details of a particular case, this needs to be
calculated in phys_grid_init.

phys_grid_init is called from within cam_initial (for each dycore
option in dynamics/XXX/inital.F90), which is itself called from
cam_init, which is called from atm_init_mct (after the call to
read_namelist). Or, phys_grid_init is called in cam_read_restart,
which is called from cam_init, in a different branch of an if test
that otherwise calls cam_initial. In both cases phys_grid_init is
called after other routines that assume that pcols has laready been
set.

This ordering of the initialization routines is not required,
and a reordering allows pcols to then be set in phys_grid_init.
This reordering is implemented here for the SE dycore.

 1) Remove ppcols from ppgrid.F90 and from phys_grid.F90.
 2) Set pcols default to -1 in ppgrid.F90.
 3) Set the pcols default to -1 in phys_grid.F90 (just to force
    model to fail until get initialization reworked) and initialize
    pcols to 16 in the body of phys_grid_init, again, for development
    purposes.
 4) Move phys_register call out of cam_init (in cam_comp.F90).
    (The call to phys_register is moved into cam_initial and into
     read_restart_dynamics. See (7) and (8) below.)
 5) Move call to phys_grid_init out of cam_read_restart (in
    cam_restart.F90). (The call to phys_grid_init for cam_read_restart
    is moved into read_restart_dynamics. See (8) below.)
 6) Move pbuf_add_field calls for FRONTGF and FRONTGA out of
    dyn_init1 (in dynamics/se/dyn_comp.F90). (These calls are moved
    into cam_initial and into read_restart_dynamics. See (7) and (8)
    below.)
 7) Add call to phys_register immediately after call to phys_grid_init
    in cam_initial (in dynamics/se/inital.F90). Add calls to
    pbuf_add_field for FRONTGF and FRONTGA immediately after the call
    to phys_register.
 8) Add calls to phys_grid_init and to phys_register (in that order)
    right after call to dyn_init1 in read_restart_dynamics (in
    dynamics/se/restart_dynamics.F90). Add calls to pbuf_add_field for
    FRONTGF and FRONTGA immediately after the call to phys_register.

[BFB]