Fixes CVMix surface layer averaging routine
Currently when cvmix computes averages for the surface layer (needed for
finding boundary layer depth) it simply does an arithmetic mean.  This
is reasonable for the 60 layer mesh (with near constant 10m resolution
through 200m), but is not correct for stretched
grids.  Here the averaging is changed to a layerThickness weighted
average.