Fix limiter calls for MMF
Fix calls to the `clip_values` routines for MMF. These were recently
changed to functions instead of subroutines (#3630) so that we can decide at the
calling level whether to handle error messages as fatal errors or
warnings, and somehow the PR that changed the behavior missed updating
the MMF radiation interface. Also adds the limiters on temperature and
optical properties that were added in PR #3630.