Eliminate unused space
When pcols_opt < 1, after columns are assigned to chunks and chunks
are assigned to processes, reset pcols for each process to the minimum
necessary, i.e. to the maximum number of columns assigned to a chunk
that is assigned to the process, constrained by phys_chnk_fdim_mult.

Also reset the gcols array within the chunk data structure to be
dimensioned ncols (instead of pcols) for each chunk, and delay the
allocation and setting of the lon and lat arrays within the
chunk data structure until after the chunks are created, so that these
can be dimensioned ncols for each chunk.

[BFB]