Change phys_chnk_fdim_max defaults
Based on empirical data, add a rule to set the default for
phys_chnk_fdim_max to 128 when running on Cori-KNL using the
SE dycore and not using the phys_grid. Also set the generic default
to be 16, mimicking the current PCOLS default.

Also change the default for phys_chnk_cost_write to .true.

[BFB]