Add memory for new crop types
This is the start of the PR to add all the new crop types in ELM as
placeholders. A new surface data and parameter file are required.
The crop model now has 34 specific crop types (and two generic crop
types).

All the below flags are true ONLY when the crop model is on!
[FCC] - Flag Climate Changing
[Non-BFB] - Non Bit-For-Bit
[CC] - Climate Changing
[NML] - Namelist Changing

See confluence for a more detailed description about these tags.