Update provenance capture logic in maint-1.0
When Compy support was added for maint-1.0, the provenance.py file
was not updated to also enable capture of provenance data for Compy.
The job monitoring script syslog.compy was also not added.

Also, when Anvil changed from a PBS batch scheduler to SLURM, the
provenance.py and syslog.anvil files in maint-1.0 were not updated
accordingly.

Finally, the projects for which performance archiving is supported
by default has not been updated in maint-1.0 for core-haswell,
cori-knl, theta, and summit.

Because of these missing changes, expected provenance data is not
being captured with the other performance data in the performance
archive. Here these deficiencies are remedied.

[BFB]