Create some utility for md arrays.
MD arrays are simply std::array<T,N>, with T potentially an std::array itself.
But the syntax
  md_array<Real,2,3,4> a;
is definitely clearer than the long and verbose std:array<...>.

Additionally, we add a data(md_array&) utility, which retrieves a pointer
to the underlying inner-most data. This is different from std::array<T,N>::data,
which returns T*, regardless of what it is (hence, for md array, T* would
be a pointer to a std::array itself). Instead, for md_array<Real,2,3,4> a,
the free function data(a) will return Real*.