Add script to setup a run directory for the atmosphere cores
This commit adds a new directory to the `testing_and_setup/` directory:
`atmosphere/` and adds a single script inside the `atmosphere/` directory. This
script, when called, prepares the current working directory for running both
the init_atmosphere and atmosphere cores by:

        * Symbolically linking both the init_atmosphere and atmosphere
        executables and, the physics lookup tables for the atmosphere core.
        These files are symbolically linked because they are both only read and
        never altered and are also quite large in size.
        * Copying all namelists, streams and stream_list files from the
        `defaults_input/` directory for each core. These files are copied
        because they will be need to be altered for each simulation.

The reason for adding this script is to automate the simple, yet redundant task
of setting up a run directory with the above files and to help new users in
setting up their run directories.

Currently, the script needs to reside in the test_and_setup/atmosphere when it
is called so it can correctly find the above files.