Regional-initialization of snowpack
For cold start simulations where use_extrasnowlayers is true,
the initialization for snowpack is set to 50 mm water eq. for
all glacier land units and vegetaed land units north of 44 degrees.
With the firn model activated, snowpack depth initialization
is not a function of the h2osno_max, which can now be increased
without affecting the initial conditions.

[NML] - Namelist Changing