Fresh snow density parameterization
Following VanKampenhout et al., 2017, I updated the new snow density
parameterization to that used by the CESM/CTSM.  This adds a wind
dependent term to the new (fresh) snow density so that as wind speed
increases, so does freshly fallen snow density.

This new parameterization is only used if use_extrasnowlayers is
true.