Fix read/write routines to allow for # mpi rank > 1 in scorpio_interface
This commit solves an issue which resulted from running the scorpio_interface
tests with the number of PIO ranks > 1.

The fix involves allows individual ranks to assign themselves a subset of the
total number of degree's of freedom for a given data array for I/O purposes.

In the case where # MPI ranks > total dof, then excess mpi ranks will skip
I/O for that array.

The scorpio_interface unit tests have been updated reflecting this change.

Additionally, substantial comments have been included in the tests in order
to facilitate future users to understand each step that goes into I/O.

Changes that were requested for PR #451, but were not incorporated before that
PR was merged by the autotester have been made in this PR.