Replace 'sqs -f' call with 'scontrol show jobid' on NERSC systems
A recent update to NERSC system software removed the '-f' option
from the sqs command. 'sqs -f' is called from provenance.py for
both cori-knl and cori-haswell, causing job submission to fail
on those systems. 'sqs -f' is also called within the job progress
monitoring scripts syslog.cori-knl and syslog.cori-haswell,
which will cause these scripts to abort (though this will not affect
a running model). 'sqs -f' is equivalent to 'scontrol show jobid',
so here 'sqs -f' is replaced by 'scontrol show jobid' in provenance.py
and in the job progress monitoring scripts. Note that the output
from 'scontrol show jobid' that is captured in provenance.py is
still saved in a file with the original name (sqsf_jobid.$lid),
because this name is looked for in the performance archiving
postprocessing scripts.

[BFB]