Update SCREAM compsets to use CMIP6 2010 tunings (#420)
* Use theta-l dycore for all SCREAM compsets

* Add 2010_scream_hr use case

* Have SCREAM-DYAMOND1 compsets run with hi-rez IFS SST

* Add 2010 use case for FSCREAM-LR

* Use semi_lagrange_cdr_alg = 3 for SCREAM

* Use se_ftype=1 for SCREAM configs

* Switch back to se_ftype = 2

Switch back to se_ftype = 2, as this should be more appropriate for
ne1024 now, and we also need it for lower resolutions and want to keep
our configurations as consistent as possible.

* Fix typo in deep_scheme

* Remove tstep_type from low-res compset (use default)

* Remove unused MG2 and CLUBB options

* Remove unused options for HR compset

* Update 2010_scream_hr.xml

set history_aerosol to false

* Update 2000_scream_lr.xml

Turned off history_aerosol and set se_ftype=2 for low res.

* set macmic_num_steps=1 in HR compset

* Switch to pg2 tests for SCREAM test suite

* Add F2000 and rename FSCREAM compsets

Adds support for the old F2000 version of the SCREAM compsets (using the
existing use case file). We now have F2000 and F2010 versions of the
SCREAM compsets, so this required renaming the compset aliases to
distinguish between the two. So, we now have the following compsets:

    * F2000-SCREAM-LR
    * F2000-SCREAM-HR
    * F2000-SCREAM-LR-DYAMOND1
    * F2000-SCREAM-HR-DYAMOND1
    * F2010-SCREAM-LR
    * F2010-SCREAM-HR
    * F2010-SCREAM-LR-DYAMOND1
    * F2010-SCREAM-HR-DYAMOND1

The F2000 versions may go away at some point, but at least for the time
being this allows us to compare with the old compsets. The new naming
convention also makes it more clear what forcing is being used.

* Rename SCREAM tests for new compsets

* Set cld_macmic_num_steps per resolution

* Rename FSCREAM-SA for consistency

Co-authored-by: crterai <rterai@uci.edu>
Co-authored-by: Peter Caldwell <caldwell19@llnl.gov>