Delay reordering of bands for consistency with optics
Delay the reordering of bands for consistency with aerosol optics
routines, which output diagnostic quanties in buried subroutines. This
fixes the issue where we were outputting the wrong band for the
diagnostic aerosol optical properties. In order to accomplish this, the
band limits in radconstants are changed to reflect the RRTMG rather than
the RRTMGP order, which is consistent with the optical property input
data and how optical properties are computed. The bands are then
reordered at the last moment to be consistent with RRTMGP absorption
coefficient data, so now throughout the working part of the code on the
E3SM side, bands should all be consistent with the old treatment used in
RRTMG and used in our current optics data. This is somewhat confusing,
but seems necessary to work with our hack of maintaining one set of
input files for RRTMG and RRTMGP. The more elegant solution might be to
stop using the RRTMG input data and create new optical properties
look-up tables consistent with RRTMGP bands (or at least include the
band information in the files so we can read and map to the expected
order).