Added capability for catch-based tests to select a kokkos device
The user must pass to the test a list of available devices, like

   ./my_test --scream-test-devices dev1[:dev2[dev3[...]]]

Scream will distribute those devices among ranks in MPI_COMM_WORLD, in a
round-robin fashion. Then, a string will be added to the argv forwarded
to kokkos, specifying the device id.

Note: the device id is set to -1 for non-cuda builds.