Initialize phys_proc_cost
The variable phys_proc_cost was introduced in a recent PR to measure
the time spent in the loops over chunks in physpkg.F90. This captures
a running total of the form

  phys_proc_cost = phys_proc_cost + <update>

after each exection of the indicated loops over chunks, and so assumes
that phys_proc_cost is initialized to 0.0_r8. However, this
initialization is missing. During testing, compilers (by default)
initialized this variable to zero, but this is not guaranteed, and
tests with DEBUG set to true in env_build.xml failed due to this
uninitialized variable. Here the required initialization is
introduced.

[BFB]