Add, initialize and deallocate mpas_kd_tree to init_atm_static
This commit imports, initializes and finalizes the mpas_kd_tree in
mpas_init_atm_static in preparation for parallelized interpolation of routines.

Currently, locations of datatile pixels are found by using the nearest_cell
routine, which finds the nearest cell from a point by randomly selecting a
cell, calculating the distance to that cell from the point and seeing if any of
the neighbors are closer to the point and if they are, recursively repeated the
processes at that point.

nearest_cell can fail when running on multiple processors if a process has a
concave decomposition, hence why multi-process is disabled for static fields.

Because finding nearest points with the KD Tree does not rely on the geometric
formations of decompositions it can replace the nearest_cell routine to
interpolate datatile pixels to cells.