Import, initialize and finalize the geotile_manager
This commit imports the mpas_geotile_manager into the mpas_init_atm_static and
initializes it and then finalizes it. Other than the initializations and
finalization of the mpas_geotile_manager, the mpas_geotile_manager is not used.

This commit is in preparation for the parallelization of the ter field.