Parallel interpolation of the ter field
This commit adds logic to allow for parallelized interpolation of the ter
(terrain) field. At present, the init_atmosphere core is still not able to
interpolate the ter field in parallel because of the check within
init_atm_setup_case of the init_atm_cases module (Lines 217-222 of
mpas_init_atm_cases.F) which is still present and prevents multiprocessing of
static fields.

Once all fields have been updated to interpolate in parallel, the above lines
in mpas_init_atm_setup_case, can be removed.