Rewrite remove_calving_ice & distribute_calving_flux to use edges
This commit rewrites the machinery that converts calvingVelocity to
actual mass removal to be baesd around edges instead of cells.  This is
a more natural way to define the calving front and allows substantial
simplification of the code.  These simplifications also eliminate a
previous issue in behavior that was dependent on the order over which cells
were looped.  This had resulted in nonuniform retreat for a uniform test
case.

Note that a number of variables have been renamed or had their
dimensions changed.

This new formulation requires that calvingVelocity be converted from
cells to edges, which has not been included in this commit.