Disable the build of the stand-alone scorpio test during CIME build.
This commit disables the building of the scorpio stand-alone test
when scream is built using CIME, as is the case with the F2000-SCREAM-SA
test.

Solves an issue with the Cori-KNL nightly tests where the SCREAM-SA test
was failing during the build step due to an error linking libraries for
the scorpio stand-alone test.  Note that we don't use this test in the
nightly testing so it isn't necessary.

[BFB]