Simulations were not BFB with history verbose turned off, now they are
Simulations are BFB now with history verbose true. "qsrflx" variable
was non-BFB which was causing history variables to differ.

It was because I was not sending info about whether the call is for
cloud borne aerosols or the interstitial aerosols.

I fixed it and now codes are BFB with ne4 and ne30 simulations (5 days)