Improve time and memory of plot_initial_state
This merge converts plot_initial_state to use xarray with chunking
to use dask parallelism.  It also uses xarray's version of
histograms.  The plotting goes for timing out or running out of
memory even for 3 hour jobs to running in 15 seconds and using
about 1% of the memory of a Grizzly compute node.