Improve bookkeeping, zap stranded ice
This commit zaps any stranded floating ice that is left after the
distribution of the caving flux.

It also improves the bookkeeping and adds a number of warnings and
errors at the end of the routine to ensure the results are reasonable.
Thresholds have been added that throw errors for situations that
probably require improvements to be made to the algorithm.

As part of the bookkeeping, all operations and totals are only done over
owned cells to avoid double counting when running on multiple
processors.

There is now an organized summary of the calving operations printed in the log at
the of the subroutine.