Additional adjustments/cleanup for grounded marine margins
Some of this fixes subtle issues, some of this is just whitespace
cleanup and reorganization I did while trying to find the subtle issues.

With this change results for grounded marine cliffs retreat at the same
rate as floating calving fronts. with one caveat.  When there is a large
thickness above flotation in the grounded marine cliff case, the calving
algorithm will first reduce the thickness there until the ice goes
afloat and then start to retreat it backwards.  This has the result of
reducing the volume calved in proportion to the reduction in thickness.
That reduction in thickness does not happen for floating calving fronts
because the thickness used for calculating the calving volume in those
situations comes from the ice thickness behind the margin cells where
the ice shelf is its 'true' thickness.  For grounded marine margins with
a small fraction above floatation, (typical Greenland glacier), this
difference is a small, acceptable error.