Removes F2000-SCREAM-SA from list of nightly test suites.
Consistent failures of this test on during the Cori-KNL
nightly tests has become a distraction.  After discussion, it has
been determined that the configuration of the SCREAM-SA test is
not consistent with the component coupler and thus failures could
be attributable to the inproper usage rather than a bug in the code.

This commit will suspend application of the F2000-SCREAM-SA nightly
test until we transition to a fully active ATM model from the
perspective of the component coupler.

[BFB]