Disable calving in MALI by default
This commit sets the calving method to 'none'.  The previous default of
having calving operating made no sense because restore_calving was also
on.

This commit also makes sure that calving will always be set to 'none'
when MALI is in STATIC mode.  For now this is redundant with the first
change, but if we change the default calving behavior in the future, we
will still want no calving to occur in STATIC mode.