Increase field widths in atm_chunk_cost.txt
The field width for the chunk id in the file atm_chunk_cost.txt is too
small when the number of chunks is 500,000 or larger, resulting in the
output of "*****". The field width for the process id in the file
is also too small when the number of processes is 1,000,000 or larger.

Here both of these conditions are checked for. If either is true, then
the field widths for chunk id and process id are increased enough
to support chunk and process counts up to 999,999,999 .

[BFB]