Removed if for listid (modal_aer_opt) and fixed bug in calcsz
Now progrnostic and diagnostic calls the calcsize and wateruptake
in the same way. The list id check is within calcsize routine.

There was a bug in the original implementation where, the last mmode
was always used in size calculations in accum<->ait transfer. That
bug is fixed in the commit along with some minor cleanup.