Update mksurfdata defaults for 0.5x0.5 res and fix a small bug
The default filenames for inputs to mksurfdata at 0.5x0.5 res
were not correct, so they have been updated.
Also, a small tolerance bug has been fixed. A pft tolerance in
normalizencheck_landuse has been updated to match its neighbor and to
match the pft tolerance in the mkpft function, which is called prior
to the call to normalizencheck_landuse.

	modified:   ../../../bld/namelist_files/namelist_defaults_clm4_5.xml
	modified:   src/mksurfdat.F90