Remove FFLAGS from definition of FC provided to esmf_time build
The src/external/Makefile defined FC as "$(FC) $(FFLAGS)" when invoking
a sub-make in the esmf_time_f90 directory. However, the Makefile in
esmf_time_f90 explicitly adds FFLAGS to compilation commands, leading to
a duplication of flags. The NAG compiler fails with the following:

  Option warning: -free option specified twice, continuing
  Option error: -convert= option specified twice

This issue is resolved by defining FC as simply "$(FC)" when invoking
the sub-make for the esmf_time target in src/external/Makefile.