Simplified Field class and surrounding datatypes.
This PR removes the Device template parameter from the Field class and its
associated types. This expresses our decision to always store Field data
on DefaultDevice (CPU for CPU-enabled calculations, and GPU for GPU-enabled
ones).

We also change the remaining template parameter from ScalarType to RealType
in the Field and related classes, to reflect the fact that Fields are always
stored as 1D arrays of real-valued numbers (and never Packs).

These two changes simplify the Field class, and hopefully make it easier to
understand the relationships between Fields and their (possible reshaped)
views.

In order to make these changes, I had to make very minor edits in the
atmosphere process types and their P3/SHOC/ZM/Radiation subclasses. Apologies
for the inconvenience.