Add and use categorical defaults to geotile manager
This commit updates the goetile_manager to add default values for categorical
parameters such as isice, iswater, isurban, etc. as specified in the section
3-53 of the WRF ARW Users' Guide.

It then updates the static interpolation of the landuse category and soiltop
category to either use the defaults, created above, or the values found in the
index file of the specified dataset of the above parameters.

Because the landuse values of isice and iswater are used to fix soil type over
Antarctica, fix inconsistent soil and landuse data, and to compute the
laskmask, their values are transfered to a non-pointer variable to be used in
the above computations. Without this change, then the above computations would
use deallocated/garbage values of isice and iswater, as the pool which
contained them is destroyed after the landuse interpolation.