Bug fix to account for snow backscattering in COSPv2
Fix a bug related to the calculation of separated ice attenuated
backscatter in lidar simulator. The issue was found by Greg Cesana.

The snow is not well accounted for in the separated ice and liquid
attenuated backscatter signal (although it's fine for the total
attenuated backscatter). It has to do with the "npart" indices.
The snow index is 5 and is not accounted for in COSPv2.  This has
potentially large impact on cloud phase diagnostics from the
lidar simulator.

[BFB] except for COSP diagnostics