SCREAM: add options to test-all-scream for quick testing
The options allow to not purge the build dir, and to not run the config step.
This allows users to do minimal changes to the code, and rerun test-all-scream,
without rebuilding the whole project.
The option --quick-rerun reruns all tests, while --quick-rerun-failed only reruns
the tests that failed in the last ctest call (if all passed, all tests are rerun).

Note: if testing against a baseline commit, baselines will still be regenerated.
      To unleash the full potential of this option, provide a baseline dir, so that
      only the test phase is executed.