SCREAM: bug fix in test-all-scream
The --quick-rerun option should skip the config step only if it was
already performed (e.g., CMakeCache.txt is found in the directory).
The presence of CMakeCache.txt is not per se enough to guarantee that
'make' will succeed when run from the bld dir (possibly automatically
running cmake again), but it is a reasonable compromise.