SCREAM: allow user to have a local mach specs file
The file must be called scream_mach_specs.py, and located in
the .cime subfolder of the user home folder (where also the
config_machine.xml and config_compiler.xml files for cime can
be placed). In order to use this file, the user needs to pass
the -l/--local flag. This flag is mutually exclusive with the
machine flag.