Update halos to 3; set thermal thickness min to 0
This commit makes two changes necessary to get correct behavior with
extended cells used by Albany.

The extension of the mesh for Albany requires 3 to cover all geometric possibilities.

The thermal minimum thickness is set to 0 to ensure reasonable values everywhere there
is ice, which is needed for realistic behavior when using Albany due to the way
temperature is handled in extended cells.