SCREAM: minor fixes to cime builds
- do not let P3 init inputs
- do not inspect atm DAG during the AD initialization
- do not call prim_printstate in cime runs

For the latter: the state is still full of zeros, so eos craps out.
I'm not sure if this is just because we're still missing some piece
in the scream-cime infrastructure, or if it should already be initialized
correctly (with nonzero values), or if this is exactly what it is supposed
to be for non-standalone runs (so no eos checks). I suspect something is
not init-ed correctly.