Enabled a year offset as an option for reading in harvest data
The harvest data for model year y are stored in year y+1 in the landuse
timeseries files. So y+1 needs to be read in
dynHarvest_interp_harvest_types(). This required an optional year offset
to be added to the function set_current_year_get_year(), in order to
utilize the optional offset in get_curr_date().