Change default location for SAVE_TIMING_DIR on Chrysalis
Currently, jobs on Anvil and Chrysalis archive performance data
in the same location. This may cause errors if Jenkins jobs,
one on Anvil and one on Chrysalis, attempt to access these
performance data simultaneously. To avoid this possibility,
the default location for archiving performance data, specified
in SAVE_TIMING_DIR, is changed for Chrysalis.

[BFB]