Fix uninitialized pblt600 and replace a remaining FC5AV1C-04P2 test
Uninitialized pblt600 could be used when dCCAPE_ULL is used and surface
pressure of the column is above 6000 hPa. It would impact how convective
adjustment is performed, which could lead to instability. It is now initialized
to 1.

Also replace a remaining test that uses the now non-existent compset FC5AV1C-04P2
with FC5AV1C-L. The test name changes but the results will not.

[BFB] except with dCAPE_ULL trigger -- as for v2 default setting -- on some systems