SCREAM: modified grids methods for layout creation
Instead of a single get_native_dof_layout method, the class
now exposes 4 methods: two for 2d and two for 3d layouts.
Those pairs consist of a scalar and a vector layout respectively.

For the vector layouts, the user can specify the tag and extent
for the vector dimension. Furthermore, the usre can choose,
via a bool, if the 3d layout is at level midpoints or interfaces.