SCREAM: add host mirror inside Field
The HM is lazy-inited, so doesn't add memory usage unless needed.
Likely, it will only be needed while porting is ongoing, and part of
the init/run sequence is still in F90.

Both get_view and get_reshaped_view now accept an extra template arg,
which is the device or mem/exec space where the view is desired. If not
provided, defaults to the device.

We also have methods to sync device and host views, but Field makes *no*
effort to keep track of whether they are in sync (it has no way to do that).