SCREAM: make Field::get_reshaped_view work for exact rank only
Before, one could use get_reshaped_view to get the underlying 1d view
with a different value type (e.g., change to packs). Now the reshape
has to be to a data type of the correct rank (as stored in the field
layout struct). If you want the 1d view, but with a different value type,
you should call get_view, and then create the new view from that one,
by using reinterpret_cast on the ptr.