Fix an issue that uses dcape related codes when ZM is not used
A block of codes in tphysac access pbuf for T_STAR and Q_STAR
that are used for dcape calculation. The block should only
be invoked when deep_scheme = 'ZM' but proper condition was not
applied, causing the test F-EAMv1-AQP1.eam-clubb_only to fail as
it sets deep_scheme = 'off'. The fix now adds a condition on
deep_scheme. The above test only became an issue after #4033 which
set default zmconv_trigdcape_ull = .true.

[BFB] for all tests, while F-EAMv1-AQP1.eam-clubb_only would report
      diff due to #4033.